Move dashboard to /hosts page
Moved the hosts stuff from the dashboard to its own seperate page.
Also added hosts to navbar and started adding the ISO section on the
host page.
