+++ app/app.go
@@ -27,6 +27,7 @@ func CreateApp() (app *fiber.App) {
 	app.Get("/login", showLogin)
 	app.Get("/logout", routesMustBeLoggedIn, showLogout)
 	app.Get("/dashboard", showDashboard)
+	app.Get("/hosts", showHosts)
 
 	// Auth API
 	app.Post("/api/auth/login", apiLogin)
