+++ public/views/layout.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>OpnLaaS - {{.Title}}</title>
+    <title>{{.Title}} - OpnLaaS</title>
     <meta name="description" content="Open Source Lab as a Service">
     <meta name="theme-color" content="#0ea5e9">
     <meta name="color-scheme" content="light dark">
@@ -28,13 +28,11 @@
 
             <nav class="nav-desktop items-center gap-6 navbar">
                 <a href="/" id="home">Home</a>
+                <a href="/dashboard" id="dashboard">Dashboard</a>
+                <a href="/hosts" id="hosts">Hosts</a>
+
                 {{if not .LoggedIn}}
                 <a href="/login" id="login">Log In</a>
-                {{end}}
-                {{if .LoggedIn}}
-                <a href="/dashboard" id="dashboard">My Dashboard</a>
-                {{else}}
-                <a href="/dashboard" id="dashboard">Public Dashboard</a>
                 {{end}}
             </nav>
 
@@ -105,24 +103,20 @@
         </div>
     </footer>
 
-    <!-- for mobile -->
-
+    <!-- Mobile Nav -->
     <nav class="md:hidden fixed bottom-0 left-0 z-50 w-full bg-background-primary/95 backdrop-blur border-t border-background-muted flex justify-around items-center px-2 py-3 pb-safe shadow-[0_-1px_3px_rgba(0,0,0,0.05)]">
-        
         <a href="/" class="flex flex-col items-center justify-center w-full h-full text-font-secondary hover:text-accent-primary focus:text-accent-primary transition-colors gap-0.5">
             <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
                 <path stroke-linecap="round" stroke-linejoin="round" d="m2.25 12 8.954-8.955c.44-.439 1.152-.439 1.591 0L21.75 12M4.5 9.75v10.125c0 .621.504 1.125 1.125 1.125H9.75v-4.875c0-.621.504-1.125 1.125-1.125h2.25c.621 0 1.125.504 1.125 1.125V21h4.125c.621 0 1.125-.504 1.125-1.125V9.75M8.25 21h8.25" />
             </svg>
             <span class="text-[10px] mt-0 font-medium">Home</span>
         </a>
-
         <a href="/dashboard" class="flex flex-col items-center justify-center w-full h-full text-font-secondary hover:text-accent-primary focus:text-accent-primary transition-colors gap-0.5">
             <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
                 <path stroke-linecap="round" stroke-linejoin="round" d="M3.75 6A2.25 2.25 0 0 1 6 3.75h2.25A2.25 2.25 0 0 1 10.5 6v2.25a2.25 2.25 0 0 1-2.25 2.25H6a2.25 2.25 0 0 1-2.25-2.25V6ZM3.75 15.75A2.25 2.25 0 0 1 6 13.5h2.25a2.25 2.25 0 0 1 2.25 2.25V18a2.25 2.25 0 0 1-2.25 2.25H6A2.25 2.25 0 0 1 3.75 18v-2.25ZM13.5 6a2.25 2.25 0 0 1 2.25-2.25H18A2.25 2.25 0 0 1 20.25 6v2.25A2.25 2.25 0 0 1 18 10.5h-2.25a2.25 2.25 0 0 1-2.25-2.25V6ZM13.5 15.75a2.25 2.25 0 0 1 2.25-2.25H18a2.25 2.25 0 0 1 2.25 2.25V18A2.25 2.25 0 0 1 18 20.25h-2.25A2.25 2.25 0 0 1 13.5 18v-2.25Z" />
             </svg>
             <span class="text-[10px] mt-0 font-medium">Dashboard</span>
         </a>
-
         {{if not .LoggedIn}}
         <a href="/login" class="flex flex-col items-center justify-center w-full h-full text-font-secondary hover:text-accent-primary focus:text-accent-primary transition-colors gap-0.5">
             <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
@@ -131,7 +125,6 @@
             <span class="text-[10px] mt-0 font-medium">Log In</span>
         </a>
         {{end}}
-
     </nav>
 </body>
